myapp/appback/requirements.txt
Django>=4.2,<5.0
# Django REST Framework
 djangorestframework>=3.14,<4.0
# Filtering
 django-filter>=23.1,<24.0
# PDF generation
 reportlab>=4.0,<5.0
# For PostgreSQL (optional, comment out if using SQLite)
# psycopg2-binary>=2.9,<3.0
# For CORS (optional, if you use it)
# django-cors-headers>=4.0,<5.0
